pyproject.toml
[project]
name = "bonds-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "textual>=0.70.0",
    "pyyaml>=6.0",
]
